+++ packages/bridge-controller/src/constants/bridge.ts
@@ -20,6 +20,7 @@ export const ALLOWED_BRIDGE_CHAIN_IDS = [
   CHAIN_IDS.BASE,
   CHAIN_IDS.SEI,
   CHAIN_IDS.MONAD,
+  CHAIN_IDS.HYPEREVM,
   SolScope.Mainnet,
   BtcScope.Mainnet,
   TrxScope.Mainnet,
@@ -62,6 +63,7 @@ export const DEFAULT_CHAIN_RANKING = [
   { chainId: 'eip155:10', name: 'Optimism' },
   { chainId: 'eip155:143', name: 'Monad' },
   { chainId: 'eip155:1329', name: 'Sei' },
+  { chainId: 'eip155:999', name: 'HyperEVM' },
   { chainId: 'eip155:324', name: 'zkSync' },
 ] as const;